+++ EssentialFeed/EssentialFeedTests/FeedItemsMapperTests.swift
@@ -1,0 +1,67 @@
+import XCTest
+import EssentialFeed
+
+final class FeedItemsMapperTests: XCTestCase {
+    
+    func test_map_throwsErrorOnNon200HTTPResponse() throws {
+        let json = makeItemsJSON([])
+        let samples = [199, 201, 300, 400, 500]
+        
+        try samples.forEach { code in
+            XCTAssertThrowsError(
+                try FeedItemsMapper.map(json, HTTPURLResponse(statusCode: code))
+            )
+        }
+    }
+    
+    func test_map_throwsErrorOnInvalidJSON() {
+        let invalidJSON = Data("invalid json".utf8)
+        
+        XCTAssertThrowsError(
+            try FeedItemsMapper.map(invalidJSON, HTTPURLResponse(statusCode: 200))
+        )
+    }
+
+    func test_map_deliversNoItemsOn200HTTPResponseWithEmptyJSON() throws {
+        let emptyJSON = makeItemsJSON([])
+        
+        let result = try FeedItemsMapper.map(emptyJSON, HTTPURLResponse(statusCode: 200))
+        
+        XCTAssertEqual(result, [])
+    }
+    
+    func test_map_deliversItemsOn200HTTPResponseWithJSONItems() throws {
+        let item1 = makeFeedImage()
+        let item2 = makeFeedImage(description: "a description", location: "a location")
+        let json = makeItemsJSON([item1.json, item2.json])
+        
+        let result = try FeedItemsMapper.map(json, HTTPURLResponse(statusCode: 200))
+        
+        XCTAssertEqual(result, [item1.model, item2.model])
+    }
+    
+    // MARK: - Helpers
+    
+    private func makeFeedImage(description: String? = nil, location: String? = nil, imageURL: URL = URL(string: "http://image-url.com")!) -> (model: FeedImage, json: [String: Any]) {
+        let model = FeedImage(id: UUID(), description: description, location: location, url: imageURL)
+        let json = [
+            "id": model.id.uuidString,
+            "description": model.description,
+            "location": model.location,
+            "image": model.url.absoluteString
+        ].compactMapValues { $0 }
+        return (model, json)
+    }
+    
+    private func makeItemsJSON(_ items: [[String: Any]]) -> Data {
+        try! JSONSerialization.data(withJSONObject: ["items": items])
+    }
+    
+}
+
+private extension HTTPURLResponse {
+    
+    convenience init(statusCode: Int) {
+        self.init(url: anyURL(), statusCode: statusCode, httpVersion: nil, headerFields: nil)!
+    }
+}
