--- EssentialFeed/EssentialFeedTests/LoadFeedFromRemoteUseCaseTests.swift
@@ -1,99 +1,0 @@
-import XCTest
-import EssentialFeed
-
-final class LoadFeedFromRemoteUseCaseTests: XCTestCase {
-    
-    func test_load_deliversErrorOnNon200HTTPResponse() {
-        let (sut, client) = makeSUT()
-        
-        let samples = [199, 201, 300, 400, 500]
-        samples.enumerated().forEach { index, code in
-            expect(sut, toCompleteWith: failure(.invalidData), when: {
-                client.complete(withStatusCode: code, data: makeItemsJSON([]), at: index)
-            })
-        }
-    }
-    
-    func test_load_deliversErrorOnInvalidJSON() {
-        let (sut, client) = makeSUT()
-        expect(sut, toCompleteWith: failure(.invalidData), when: {
-            let invalidJSON = Data("invalid json".utf8)
-            client.complete(withStatusCode: 200, data: invalidJSON)
-        })
-    }
-
-    func test_load_deliversNoItemsOn200HTTPResponseWithEmptyJSON() {
-        let (sut, client) = makeSUT()
-        expect(sut, toCompleteWith: .success([])) {
-            let emptyJSON = makeItemsJSON([])
-            client.complete(withStatusCode: 200, data: emptyJSON)
-        }
-    }
-    
-    func test_load_deliversItemsOn200HTTPResponseWithJSONItems() {
-        let (sut, client) = makeSUT()
-
-        let item1 = makeFeedImage()
-        let item2 = makeFeedImage(description: "a description", location: "a location")
-        
-        expect(sut, toCompleteWith: .success([item1.model, item2.model])) {
-            let json = makeItemsJSON([item1.json, item2.json])
-            client.complete(withStatusCode: 200, data: json)
-        }
-    }
-    
-    // MARK: - Helpers
-    
-    private func makeSUT(url: URL = URL(string: "http://a-url.com/")!, file: StaticString = #filePath, line: UInt = #line) -> (sut: RemoteFeedLoader, client: HTTPClientSpy) {
-        let client = HTTPClientSpy()
-        let sut = RemoteFeedLoader(client: client, url: url)
-        
-        trackForMemoryLeaks(sut)
-        trackForMemoryLeaks(client)
-        
-        return (sut, client)
-    }
-    
-    private func failure(_ error: RemoteFeedLoader.Error) -> RemoteFeedLoader.Result {
-        return .failure(error)
-    }
-    
-    private func makeFeedImage(description: String? = nil, location: String? = nil, imageURL: URL = URL(string: "http://image-url.com")!) -> (model: FeedImage, json: [String: Any]) {
-        let model = FeedImage(id: UUID(), description: description, location: location, url: imageURL)
-        let json = [
-            "id": model.id.uuidString,
-            "description": model.description,
-            "location": model.location,
-            "image": model.url.absoluteString
-        ].compactMapValues { $0 }
-        return (model, json)
-    }
-    
-    private func makeItemsJSON(_ items: [[String: Any]]) -> Data {
-        try! JSONSerialization.data(withJSONObject: ["items": items])
-    }
-    
-    private func expect(_ sut: RemoteFeedLoader, toCompleteWith expectedResult: RemoteFeedLoader.Result, when action: () -> (), file: StaticString = #filePath, line: UInt = #line) {
-        
-        let exp = expectation(description: "Wait for load completion")
-        
-        sut.load { receivedResult in
-            
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedItems), .success(expectedItems)):
-                XCTAssertEqual(receivedItems, expectedItems, file: file, line: line)
-            case let (.failure(receivedError as RemoteFeedLoader.Error), .failure(expectedError as RemoteFeedLoader.Error)):
-                XCTAssertEqual(receivedError, expectedError, file: file, line: line)
-            default:
-                XCTFail("Expected result \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-            }
-            
-            exp.fulfill()
-        }
-        
-        action()
-        
-        wait(for: [exp], timeout: 1.0)
-    }
-    
-}
